Extract DescriptionFormatter from EventResolverBase
